Rename the `customTab` package to `web`
