+++ sdk-web/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reach5.identity.sdk.customTab">
+          package="com.reach5.identity.sdk.web">
 
     <application>
         <activity
-                android:name=".ReachFiveLoginActivity"
+                android:name="com.reach5.identity.sdk.web.ReachFiveLoginActivity"
                 android:launchMode="singleTask">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW"/>
